Convert mypy test to script based test
