+++ test/sanity/code-smell/mypy/ansible-core.ini
