+++ test/sanity/code-smell/mypy/ansible-test.ini
