+++ test/sanity/code-smell/mypy/modules.ini
