+++ test/sanity/code-smell/mypy/packaging.ini
